create reactionList, run prettier
