+++ examples/website-demos/social-messaging/src/assets/RightReply.tsx
@@ -1,5 +1,12 @@
 export const RightReply: React.FC = () => (
-  <svg width="17" height="33" viewBox="0 0 17 33" fill="none" xmlns="http://www.w3.org/2000/svg" style={{ position: 'relative', bottom: '6px' }}>
-    <path d="M16.5 32V32C7.66344 32 0.5 24.8366 0.5 16L0.5 0" stroke="#DBDDE1"/>
+  <svg
+    width='17'
+    height='33'
+    viewBox='0 0 17 33'
+    fill='none'
+    xmlns='http://www.w3.org/2000/svg'
+    style={{ position: 'relative', bottom: '6px' }}
+  >
+    <path d='M16.5 32V32C7.66344 32 0.5 24.8366 0.5 16L0.5 0' stroke='#DBDDE1' />
   </svg>
 );
