+++ examples/website-demos/social-messaging/src/components/Message/SocialMessageUI.scss
@@ -35,6 +35,7 @@
   &-inner {
     align-items: flex-end;
     position: relative;
+    min-width: 102px;
 
     .str-chat__message-text {
       align-self: flex-start;
@@ -49,7 +50,6 @@
 
       div.my-message {
         align-self: flex-end;
-        // background: red;
       }
 
       .message-wrapper-inner-options {
@@ -132,112 +132,4 @@ div.message-wrapper-inner .str-chat__message-attachment {
   border-radius: 18px 18px 0px 18px;
   height: 36px;
   justify-content: center;
-}
-
-// reactions
-.message-wrapper-inner-options {
-  flex-direction: row-reverse;
-}
-
-.message-wrapper-inner:hover .str-chat__message-simple__actions {
-  svg:hover {
-    fill: var(--accent-primary);
-  }
-
-  &__action--reactions {
-    display: flex;
-    position: relative;
-  }
-
-  &__action--options {
-    display: flex;
-  }
-}
-
-div.str-chat__reaction-selector {
-  background: var(--white);
-  height: 40px;
-  bottom: -45px;
-  left: 60px;
-
-  .emoji-mart-emoji:hover {
-    transform: none;
-  }
-
-  ul {
-    margin: 0 8px;
-  }
-}
-
-.message-wrapper-inner.my-message.my-message > div.str-chat__reaction-selector {
-  right: 40px;
-  left: auto;
-}
-
-li.str-chat__message-reactions-list-item {
-  padding: 8px;
-  margin: 0;
-
-  &:hover {
-    background: var(--input-bg);
-    border-radius: 20px;
-  }
-}
-
-span.str-chat__message-reactions-list-item__count {
-  color: var(--black);
-}
-
-.message-wrapper-inner-text {
-  position: relative;
-}
-
-ul.str-chat__simple-reactions-list {
-  position: absolute;
-  top: -32px;
-  left: -15px;
-  // left: -55px;
-  z-index: 2;
-  border: 1px solid var(--borders);
-  height: 32px;
-  min-width: 32px;
-  border-radius: 32px;
-  align-items: center;
-  justify-content: center;
-  background: var(--app-bg);
-  box-shadow: 2px 2px var(--app-bg);
-}
-
-li.str-chat__simple-reactions-list-item {
-  padding-left: 4px;
-}
-
-.bubble-1 {
-  position: absolute;
-  height: 8px;
-  width: 8px;
-  background: var(--app-bg);
-  top: 2px;
-  left: 0;
-  z-index: 1;
-  border: 1px solid var(--borders);
-  border-radius: 8px;
-  box-shadow: 1px 1px var(--app-bg);
-}
-
-.bubble-2 {
-  position: absolute;
-  height: 4px;
-  width: 4px;
-  background: var(--app-bg);
-  top: 5px;
-  left: 5px;
-  z-index: 2;
-  border: 1px solid var(--borders);
-  border-radius: 4px;
-  box-shadow: 1px 1px var(--app-bg);
-}
-
-li.str-chat__simple-reactions-list-item--last-number {
-  display: none;
 }
